ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
boto3==1.35.0
Pillow==10.4.0
numpy==1.26.4
scipy==1.14.0
opencv-python-headless==4.10.0.84
torch
torchvision
transformers>=4.45.0
accelerate
simple-lama-inpainting
